% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.cyclo.pos.R
\encoding{UTF-8}
\name{cluster.cyclo.pos}
\alias{cluster.cyclo.pos}
\title{Clusters' cyclical start, end, mean and range}
\usage{
cluster.cyclo.pos(x, k)
}
\arguments{
\item{x}{A cyclical-time-constrained clustering object, as provided by function \code{\link{cyclo.hclust}}.}

\item{k}{an integer scalar with the desired number of clusters (as in \code{\link{cutree}}).}
}
\description{
Function \code{cluster.cyclo.pos} computes the start, end, mean and range of cyclical-time-constrained clusters along the time series cycles.
}
\seealso{
\code{\link{cyclo.hclust}}
}
\author{
Nicolas Djeghri, UBO
}
