% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclo.hclust.R
\encoding{UTF-8}
\name{cyclo.hclust}
\alias{cyclo.hclust}
\title{Cyclical time-Constrained Clustering}
\usage{
cyclo.hclust(
  d,
  method = "ward.D2",
  time,
  cycle.duration,
  cyclic.link.tolerance = cycle.duration/10
)
}
\arguments{
\item{d}{a class \code{\link{dist}} dissimilarity matrix.}

\item{method}{the agglomeration method to be used (parsed to \code{\link[adespatial]{constr.hclust}}).}

\item{time}{a vector indicating the linear time corresponding to objects in \code{d}.}

\item{cycle.duration}{a number indicating the duration of one cycle. Must be in the same units as \code{time}.}

\item{cyclic.link.tolerance}{a number indicating how far in cyclic time a link can reach (see details). Defaults to \code{cycle.duration/10}.}
}
\value{
\code{cyclo.hclust} returns an object of class \code{cyclo.hclust}. It is a
modified  version of the output of function \code{\link[adespatial]{constr.hclust}}
containing additional information on linear and cyclical time
used in the other functions of package \code{CycloClust}.
}
\description{
Function \code{cyclo.hclust} performs cyclical-time-constrained agglomerative
clustering from a multivariate dissimilarity matrix. The function piggy-backs
largely on function \code{\link[adespatial]{constr.hclust}} by P. Legendre
and G. Guénard, from package \code{adespatial}, the user
is strongly advised to check the corresponding documentation as well as
documentation from \code{\link{hclust}}, the more general R function for
hierarchical clustering.
}
\details{
Function \code{cyclo.hclust} uses the parameters \code{time},
\code{cycle.duration}, and \code{cyclic.link.tolerance} to compute a list of
edges connecting observation along linear time AND cyclical time (for
instance with a multi-annual time-series, connecting January of year 1 to
January of year 2 etc.). This list of edges is then transferred to function
\code{\link[adespatial]{constr.hclust}} as its \code{link} parameter. All the
clustering is carried by \code{\link[adespatial]{constr.hclust}} and users
are strongly advised to check the associated documentation for clustering methods.

The links generated by \code{cyclo.hclust} can be divided in two categories.
First, linear-time links simply link points contiguous in time (as in classic
chronological clustering). Second, cyclical-time links are established: For
each point, the function establishes a link with the point closest in
cyclical time (within the margin of \code{cyclic.link.tolerance}) on the next
cycle. For irregular sampling or in case of missing data, this implies that
some cyclical-time links may not be established. This may be alleviated by
increasing \code{cyclic.link.tolerance} but this implies a deformation of the
cyclical constraint.
}
\examples{
#I have to put some R code in there!!!

}
\references{
Guénard, G. and P. Legendre. 2022. Hierarchical clustering with contiguity
constraint in R. Journal of Statistical Software 103(7): 1-26
(https://doi.org/10.18637/jss.v103.i07)
}
\seealso{
\code{\link[adespatial]{constr.hclust}}, \code{\link{hclust}}
}
\author{
Nicolas Djeghri, UBO
}
