% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cyclo.clust.R
\encoding{UTF-8}
\name{plot.cyclo.clust}
\alias{plot.cyclo.clust}
\title{Clusters' cyclical start, end, mean and range}
\usage{
\method{plot}{cyclo.clust}(
  x,
  k,
  clust.colors = NULL,
  cyclical.factor = 1,
  add.means = F,
  lines.means = F,
  add.points = F,
  ...
)
}
\arguments{
\item{x}{a cyclical-time-constrained clustering object, as provided by function \code{\link{cyclo.hclust}}.}

\item{k}{an integer scalar with the desired number of clusters (as in \code{\link{cutree}}).}

\item{clust.colors}{a vector of colors of the same length as \code{k}. Defaults to shades of grey if not provided.}

\item{cyclical.factor}{a factor allowing to convert from the unit used in linear time to the unit used in cyclical time used for display.}

\item{add.means}{boolean. Should the cluster mean positions be displayed?}

\item{lines.means}{boolean. Should the means by joined by lines?}

\item{add.points}{boolean. Should the raw data points be displayed?}

\item{...}{other parameters parsed to function \code{\link{plot}}.}
}
\description{
Function \code{plot.cyclo.clust} provides options to visualize the cyclical-
time constrained clusters produced by \code{\link{cyclo.hclust}}.
}
\examples{
#WRITE SOME R CODE IN THERE!!!

}
\seealso{
\code{\link{cyclo.hclust}}, \code{\link{cluster.cyclo.pos}}
}
\author{
Nicolas Djeghri, UBO
}
