% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.cyclo.pos.R
\name{trend.cyclo.pos}
\alias{trend.cyclo.pos}
\title{Trend tests for clusters' cyclical start, end, mean and range}
\usage{
trend.cyclo.pos(x)
}
\value{
An object of class \code{\link{list}}, with four tables giving the results of
the Mann-Kendall tests (tau statistic and p.value) for the different cyclical
position indices:
\itemize{
\item{\code{means}}
\item{\code{starts}}
\item{\code{ends}}
\item{\code{ranges}}
}
}
\description{
Function \code{trend.cyclo.pos} tests for linear trends in the clusters'
cyclical position indices obtained by \code{\link{cluster.cyclo.pos}}.
}
\details{
The function performs a Mann-Kendall (using function \code{\link{cor.test}})
trend test on the different clusters' cyclical position indices obtained by
function \code{\link{cluster.cyclo.pos}}.
}
\seealso{
\code{\link{cluster.cyclo.pos}}, \code{\link{cor.test}}
}
\author{
Nicolas Djeghri, UBO
}
