% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.CycloClust.R
\encoding{UTF-8}
\name{utilities.CycloClust}
\alias{utilities.CycloClust}
\alias{circ.mean}
\alias{linearize.circ.ts}
\title{Utility functions for CycloClust}
\usage{
circ.mean(x, na.rm = T)

linearize.circ.ts(ts, cycle.duration, start.is.0 = T)
}
\arguments{
\item{x}{A vector containing angles (in radians) to be averaged}

\item{na.rm}{Boolean. Should the \code{NA} be removed prior to circular mean computation?}

\item{ts}{A circular time series that is to be linearized.}

\item{cycle.duration}{The duration of a cycle in ts}

\item{start.is.0}{Boolean. Should the start of the linearized time series be set to 0? Else it is taken as the first value of \code{ts}.}
}
\description{
Function \code{circ.mean} computes circular means while function \code{linearize.circ.ts} "linearizes" a circular time series (see details).
This is particularly useful to test trends along time (see examples).
}
\details{
Function \code{circ.mean} computes circular means.

Function \code{linearize.circ.ts} "linearizes" cyclical time series (i.e. time series of dates along cycles).
}
\references{
NONE for now!
}
\seealso{
\code{\link{cyclo.hclust}},\code{\link{cluster.cyclo.pos}}
}
\author{
Nicolas Djeghri, UBO
}
