% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.cyclo.pos.R
\encoding{UTF-8}
\name{cluster.cyclo.pos}
\alias{cluster.cyclo.pos}
\title{Clusters' cyclical start, end, mean and range}
\usage{
cluster.cyclo.pos(x, k)
}
\arguments{
\item{x}{a cyclical-time-constrained clustering object, as provided by function \code{\link{cyclo.hclust}}.}

\item{k}{an integer scalar with the desired number of clusters (as in \code{\link{cutree}}).}
}
\value{
An object of class \code{\link{list}}, containing:
\itemize{
\item{\code{means}: A table containing the mean position of each cluster in each cycle.}
\item{\code{starts}: A table containing the starting position of each cluster along each cycle.}
\item{\code{ends}: A table containing the ending position of each cluster along each cycle.}
\item{\code{ranges}: A table containing the ranges covered by each cluster within each cycle.}
\item{\code{cycle.duration}: The duration of a cycle.}
}
}
\description{
Function \code{cluster.cyclo.pos} computes indices of cluster position
(start, end, mean and range) of cyclical-time-constrained clusters along
different cycles.
}
\details{
A valuable aspect of a cyclical-time constrained clustering is the ability to
compare clusters' position in several cycles. Displacement of clusters can be
interpreted as advancing or delaying of a particular state along the cycles.
The function \code{cluster.cyclo.pos} compute four indices allowing related to
cluster position within cycles (see Value).
}
\examples{
#WRITE SOME R CODE IN THERE!!!

}
\seealso{
\code{\link{cyclo.hclust}}
}
\author{
Nicolas Djeghri, UBO
}
