% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CycloClust-package.R
\docType{package}
\name{CycloClust-package}
\alias{CycloClust-package}
\alias{CycloClust}
\title{CycloClust: Cyclical-time constrained clustering}
\description{
Constrained clustering of time series with cyclical dynamics (seasonality, day-night cycles etc.).
}
\seealso{
\link{\code{adespatial}}
}
\author{
Nicolas Djeghri
\email{nicolas.djeghri@gmail.com}
\href{https://orcid.org/0000-0001-5740-3386}{ORCID}
}
\keyword{internal}
