% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CycloClust-package.R
\docType{package}
\name{CycloClust-package}
\alias{CycloClust-package}
\alias{CycloClust}
\title{CycloClust: Cyclical-time constrained clustering}
\description{
Constrained clustering of time series with cyclical dynamics
(seasonality, day-night cycles etc.).
}
\details{
The main function of the package is \code{\link{cyclo.hclust}}, derived from
\code{\link[adespatial]{constr.hclust}}, and allowing to perform cyclical-time
constrained hierarchical clustering. Once a clustering is obtained,
\code{CycloClust} provides functions to follow changes in cluster position
within cycles (see \code{\link{cluster.cyclo.pos}}, \code{\link{lin.cyclo.ts}},
\code{\link{trend.cyclo.pos}}) and to visualize clusters (A RAJOUTER!!!).
}
\seealso{
\code{\link[adespatial]{constr.hclust}}, \code{\link{hclust}}
}
\author{
Nicolas Djeghri
\email{nicolas.djeghri@gmail.com}
\href{https://orcid.org/0000-0001-5740-3386}{ORCID}
}
\keyword{internal}
