% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CycloClust-package.R
\docType{package}
\name{CycloClust-package}
\alias{CycloClust}
\alias{CycloClust-package}
\title{CycloClust: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NicoDjegh/CycloClust}
  \item Report bugs at \url{https://github.com/NicoDjegh/CycloClust/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Djeghri \email{nicolas.djeghri@gmail.com} (\href{https://orcid.org/0000-0001-5740-3386}{ORCID}) [copyright holder]

}
\keyword{internal}
